requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
docopt==0.6.2
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
plumbum==1.6.7
psutil==5.4.7
pylint==2.1.1
python-dotenv==0.9.1
requests==2.19.1
rpyc==4.0.2
six==1.11.0
thingspeak==0.4.1
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
